requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
psycopg[binary]>=3.2,<3.3
